added new omnidirectional camera model base don Scaramuzza's camera toolbox fixed warnings for c++11 support and just a lot of other warnings

+++ src/location_recognition/LocationRecognition.cc
@@ -90,7 +90,7 @@ LocationRecognition::knnMatch(const FrameConstPtr& frame, int k,
 
         matches.push_back(tag);
 
-        if (matches.size() == k)
+        if ((int)matches.size() == k)
         {
             return;
         }
@@ -127,7 +127,7 @@ LocationRecognition::knnMatch(const FrameConstPtr& frame, int k,
 
         matches.push_back(frame);
 
-        if (matches.size() == k)
+        if ((int)matches.size() == k)
         {
             return;
         }
